Separated .h and .cpp files to avoid build errors due to duplicated includes
